requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
dnspython==2.4.2
APScheduler==3.10.4
python-dotenv==1.0.0